+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.0.2
+
+fix: `buffer.toString` now returns `''` if the buffer is empty.
+feat: added `buffer.isEmpty`
+
 ## 0.0.1
 
 Initial release